config: Add config flag for memory profiling
Adds a config flag to enable/disable the memory sampling.

Flag can be changed without a restart. However, turning it on after the
fact is not that useful because we will be missing lots of live
allocations.

It's still useful if profiling needs to be turned off for whatever
reason.