scrubber: implement deletion of orphan segments
This is very conservative: it will only eliminate
a segment if the topic has been deleted _and_ there
is a remote lifecycle marker indicating that the
topic purge is complete.

This means it won't clean up orphan segments from
older verisons of Redpanda, but that is rather hard
to do safely, as we don't have a robust record of
whether a topic was intentionally deleted.